Scroll the page on scrub rather than hide/show previous items
This allows the user to scroll back up if using the scrubber to
navigate to a different area of the page.

At the moment, we tie the progress bar and scrub marker together.
It's possible we'll want to ditch the scrub marker entirely, or
figure out what it's really supposed to look like. I guess it'd
be nice to have a drag interface associated with it as well.